shared helm pkg for cli install and bootstrapper
