+++ cli/internal/helm/BUILD.bazel
@@ -7,7 +7,6 @@ go_library(
         "backup.go",
         "client.go",
         "helm.go",
-        "installer.go",
         "loader.go",
         "symwalk.go",
         "values.go",
@@ -405,7 +404,6 @@ go_library(
         "//internal/deploy/helm",
         "//internal/file",
         "//internal/logger",
-        "//internal/retry",
         "//internal/semver",
         "//internal/versions",
         "@com_github_pkg_errors//:errors",
@@ -413,9 +411,7 @@ go_library(
         "@io_k8s_apiextensions_apiserver//pkg/apis/apiextensions/v1:apiextensions",
         "@io_k8s_apimachinery//pkg/apis/meta/v1/unstructured",
         "@io_k8s_apimachinery//pkg/runtime/schema",
-        "@io_k8s_apimachinery//pkg/util/wait",
         "@io_k8s_sigs_yaml//:yaml",
-        "@org_uber_go_zap//:zap",
         "@sh_helm_helm//pkg/ignore",
         "@sh_helm_helm_v3//pkg/action",
         "@sh_helm_helm_v3//pkg/chart",
